+++ src/main/java/com/contextualmocker/AnyMatcher.java
@@ -1,0 +1,8 @@
+package com.contextualmocker;
+
+class AnyMatcher<T> implements ArgumentMatcher<T> {
+    @Override
+    public boolean matches(Object argument) {
+        return true;
+    }
+}