+++ src/main/java/com/contextualmocker/ArgumentMatcher.java
@@ -1,0 +1,5 @@
+package com.contextualmocker;
+
+public interface ArgumentMatcher<T> {
+    boolean matches(Object argument);
+}