+++ src/main/java/com/contextualmocker/ArgumentMatchers.java
@@ -10,30 +10,48 @@ public final class ArgumentMatchers {
         return null;
     }
 
+    public static int anyInt() {
+        MatcherContext.addMatcher(new AnyMatcher<>());
+        return 0;
+    }
+
+    public static long anyLong() {
+        MatcherContext.addMatcher(new AnyMatcher<>());
+        return 0L;
+    }
+
+    public static double anyDouble() {
+        MatcherContext.addMatcher(new AnyMatcher<>());
+        return 0.0d;
+    }
+
+    public static float anyFloat() {
+        MatcherContext.addMatcher(new AnyMatcher<>());
+        return 0.0f;
+    }
+
+    public static boolean anyBoolean() {
+        MatcherContext.addMatcher(new AnyMatcher<>());
+        return false;
+    }
+
+    public static byte anyByte() {
+        MatcherContext.addMatcher(new AnyMatcher<>());
+        return (byte) 0;
+    }
+
+    public static short anyShort() {
+        MatcherContext.addMatcher(new AnyMatcher<>());
+        return (short) 0;
+    }
+
+    public static char anyChar() {
+        MatcherContext.addMatcher(new AnyMatcher<>());
+        return '\u0000';
+    }
+
     public static <T> T eq(T value) {
-        MatcherContext.addMatcher(new EqMatcher<>(value));
-        return value;
-    }
-
-    public interface ArgumentMatcher<T> {
-        boolean matches(T argument);
-    }
-
-    static class AnyMatcher<T> implements ArgumentMatcher<T> {
-        @Override
-        public boolean matches(T argument) {
-            return true;
-        }
-    }
-
-    static class EqMatcher<T> implements ArgumentMatcher<T> {
-        private final T expected;
-        EqMatcher(T expected) {
-            this.expected = expected;
-        }
-        @Override
-        public boolean matches(T argument) {
-            return Objects.equals(expected, argument);
-        }
-    }
+    MatcherContext.addMatcher(new EqMatcher<>(value));
+    return value;
+}
 }