+++ src/main/java/com/contextualmocker/MatcherContext.java
@@ -4,15 +4,15 @@ import java.util.ArrayList;
 import java.util.List;
 
 final class MatcherContext {
-    private static final ThreadLocal<List<ArgumentMatchers.ArgumentMatcher<?>>> matchers =
+    private static final ThreadLocal<List<ArgumentMatcher<?>>> matchers =
             ThreadLocal.withInitial(ArrayList::new);
 
-    static void addMatcher(ArgumentMatchers.ArgumentMatcher<?> matcher) {
+    static void addMatcher(ArgumentMatcher<?> matcher) {
         matchers.get().add(matcher);
     }
 
-    static List<ArgumentMatchers.ArgumentMatcher<?>> consumeMatchers() {
-        List<ArgumentMatchers.ArgumentMatcher<?>> current = new ArrayList<>(matchers.get());
+    static List<ArgumentMatcher<?>> consumeMatchers() {
+        List<ArgumentMatcher<?>> current = new ArrayList<>(matchers.get());
         matchers.get().clear();
         return current;
     }